+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog for the Birthdays plugin
 
 ## Version 1.0.0
-Release TBD
+Release 2020-08-31
 - Remove html-format apostrophy in dscp language string
 - Show only allowed days when birthday form is first displayed
 - Add privacy export function